backend/requirements.txt
fastapi
sqlalchemy[asyncio]
asyncpg
python-dotenv
pydantic[email]
uvicorn
celery
pydantic-settings
redis
flower
scikit-learn
numpy
pandas>=2.2.0
alembic
bcrypt
geoalchemy2
shapely
rasterio
httpx
python-jose[cryptography]
stripe